Added Keycap STL File
Here is the keycap that i modded from the opensource Adafruit cherry mx designs. I call this custom profile "DCZSA" since it was roughly based off a "DSA" profile cap with the top chopped off to aid in printing. Rotate the keycap so that the top is down on your print bed, this aids in adhesion when printing. Depending on your filament, you may need to scale up the model in 0.5mm increments to get the stem to slide on and off a keyswitch easier.